services/nutrition-facts-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pandas==2.1.4
numpy==1.25.2
pyodbc==4.0.39
python-dotenv==1.0.0
openai==1.51.0
httpx==0.27.0
